+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docs rename variable `rosterArray ` from the attendees list to `attendees`
 - Add base styles to ChannelList
 - Fixed createChatBubbleList unit test that was incorrectly named and not running
-- Fixed TS error inside /InfiniteList/index.tsx.
+- Fixed TS error inside /InfiniteList/index.tsx
+- Reselect audio input when device IDs go stale
 
 ### Added
 - Add `tabIndex` to BaseProps